feat(day10): tích hợp Swagger (Springdoc OpenAPI) và cấu hình bảo mật JWT
- Thêm dependency `springdoc-openapi-starter-webmvc-ui` vào pom.xml
- Tạo class SwaggerConfig để cấu hình OpenAPI:
  + Khai báo title, version, description cho API
  + Định nghĩa security scheme (Bearer Authentication - JWT)
- Cập nhật SecurityConfig: mở quyền cho swagger-ui và api-docs
- Refactor ProductController, AuthController, UserController:
  + Dùng @Operation để mô tả API
  + Dùng @SecurityRequirement để đánh dấu API cần JWT
  + Phân loại rõ API public (không khóa) và secured (có khóa)
- Thêm annotation @Tag cho từng controller để đổi tên hiển thị trên Swagger UI
  + ProductController → "Sản phẩm"
  + AuthController → "Xác thực"
  + UserController → "Người dùng"
  + HelloController → "Hello API"

+++ src/main/java/com/shop/onlineshop/repository/RoleRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface RoleRepository extends JpaRepository<RoleEntity, Integer> {
+public interface RoleRepository extends JpaRepository<RoleEntity, Long> {
     Optional<RoleEntity> findByName(String name);
 }
